+++ packages/react-components/react-motions-preview/src/index.ts
@@ -1,6 +1,8 @@
 import * as atoms from './atoms';
+import * as transitions from './transitions';
 
 export { createAtom } from './factories/createAtom';
+export { createTransition } from './factories/createTransition';
 
-export { atoms };
-export type { MotionAtom } from './types';
+export { atoms, transitions };
+export type { MotionAtom, MotionTransition } from './types';