+++ packages/react-components/react-motions-preview/src/types.ts
@@ -2,3 +2,8 @@ export type MotionAtom = {
   keyframes: Keyframe[];
   options: KeyframeEffectOptions;
 };
+
+export type MotionTransition = {
+  enter: MotionAtom;
+  exit: MotionAtom;
+};