+++ packages/react-components/react-teaching-popover-preview/src/components/TeachingPopoverSurface/renderTeachingPopoverSurface.tsx
@@ -1,0 +1,15 @@
+/** @jsxRuntime automatic */
+/** @jsxImportSource @fluentui/react-jsx-runtime */
+
+import { assertSlots } from '@fluentui/react-utilities';
+import type { TeachingPopoverSurfaceState, TeachingPopoverSurfaceSlots } from './TeachingPopoverSurface.types';
+
+/**
+ * Render the final JSX of TeachingPopoverSurface
+ */
+export const renderTeachingPopoverSurface_unstable = (state: TeachingPopoverSurfaceState) => {
+  assertSlots<TeachingPopoverSurfaceSlots>(state);
+
+  // TODO Add additional slots in the appropriate place
+  return <state.root />;
+};